requirements.txt
# 卫星通信干扰分析系统微服务依赖

# Web框架
Flask==2.3.3

# HTTP客户端
requests==2.31.0

# 数值计算
numpy==1.24.3

# 数据处理
pandas==2.0.3

# 绘图
matplotlib==3.7.2

# 日期时间处理
python-dateutil==2.8.2

# 多智能体报告生成服务依赖
openai==1.3.0
pillow==10.0.1
pypandoc==1.11
autogen-agentchat==0.2.0

# JSON处理 (Python内置，无需安装)
# json

# 数学计算 (Python内置，无需安装)
# math

# 操作系统接口 (Python内置，无需安装)
# os

# 系统时间 (Python内置，无需安装)
# datetime

# 命令行参数 (Python内置，无需安装)
# sys
